+++ lib/src/core_tokens/effect_primitive.dart
@@ -1,0 +1,34 @@
+/*
+ * SPDX-FileCopyrightText: Copyright 2026 LG Electronics Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import 'package:flutter/material.dart';
+
+class EffectPrimitive {
+  EffectPrimitive._();
+
+  static EffectPrimitive? _instance;
+  static EffectPrimitive get instance => _instance ??= EffectPrimitive._();
+
+  late final int effect0 = 0;
+  late final int effect1 = 1;
+  late final int effect2 = 2;
+  late final int effect3 = 3;
+  late final int effect4 = 4;
+  late final int effect5 = 5;
+  late final int effect6 = 6;
+  late final int effect7 = 7;
+  late final int effect8 = 8;
+  late final int effect9 = 9;
+  late final int effect10 = 10;
+  late final int effect12 = 12;
+  late final int effect14 = 14;
+  late final int effect16 = 16;
+  late final int effect18 = 18;
+  late final int effect20 = 20;
+  late final int effect40 = 40;
+  late final int effect60 = 60;
+  late final int effect80 = 80;
+  late final int effect100 = 100;
+}