+++ lib/src/core_tokens/opacity_primitive.dart
@@ -1,0 +1,24 @@
+/*
+ * SPDX-FileCopyrightText: Copyright 2026 LG Electronics Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import 'package:flutter/material.dart';
+
+class OpacityPrimitive {
+  OpacityPrimitive._();
+
+  static OpacityPrimitive? _instance;
+  static OpacityPrimitive get instance => _instance ??= OpacityPrimitive._();
+
+  late final double opacityA0 = 0;
+  late final double opacityA10 = 0.1;
+  late final double opacityA20 = 0.2;
+  late final double opacityA30 = 0.3;
+  late final double opacityA40 = 0.4;
+  late final double opacityA50 = 0.5;
+  late final double opacityA60 = 0.6;
+  late final double opacityA70 = 0.7;
+  late final double opacityA80 = 0.8;
+  late final double opacityA90 = 0.9;
+}