+++ packages/core-tokens/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The following is a curated list of changes in the core tokens module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `color-primitive` values to match the latest guide
+
+### Added
+
+- `effect-primitive` values to match the latest guide
+- `opacity-primitive` values to match the latest guide
+- `radius-primitive` values to match the latest guide
+- `spacing-primitive` values to match the latest guide
+- `typography-primitive` values to match the latest guide
+
 ## [2.0.0] - 2026-04-30
 
 - No significant changes